Aiming at factoring out dibelsParsing
