+++ app/assets/javascripts/reader_profile_january/LetterNamingFluencyTab.js
@@ -1,10 +1,29 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import tabProptypes from './tabPropTypes';
-import {NoInformation} from './Tabs';
+import {DIBELS_LNF} from '../reading/thresholds';
+import {mostRecentDataPoint, shouldHighlight} from './dibelsParsing';
+import {Tab} from './Tabs';
+
 
 export default class LetterNamingFluencyTab extends React.Component {
   render() {
-    return <NoInformation />;
+    const {nowFn} = this.context;
+    const {onClick, student, readerJson} = this.props;
+    
+    const dataPoint = mostRecentDataPoint(readerJson, DIBELS_LNF);    
+    if (!dataPoint) return null;
+    const isOrange = shouldHighlight(dataPoint, student.grade, nowFn());
+    return (
+      <Tab
+        text="Letter naming fluency"
+        orange={isOrange}
+        onClick={onClick}
+      />
+    );
   }
 }
 LetterNamingFluencyTab.propTypes = tabProptypes;
+LetterNamingFluencyTab.contextTypes = {
+  nowFn: PropTypes.func.isRequired
+};
